import modified version of quassel as kuassel
the version imported is based on upstream 9c5fd60 with its bundled
copies of miniz (alternative to zlib), oxygen icons and other
non-relevant cruft removed. KDE integreation hardcoded having in
mind it will improve (icons handling, browser integration, etc.)
rebranding is not done yet and credits to the original authors and
contributors are and will be left mostly intact with the exception
of adding new.