kuassel: implement standard KDE actions in help menu
this had to be done the hardway, at least for now. altough
KHelpMenu can return KMenu pointer it does not play well with the
rest of the code and will require quite some changes (AFAICT) so
for now I opted to re-create standard KHelpMenu with the actions
available for most applications (see TODO bellow). sadly, whatsThis
slot is not yet part of the public KDE UI API (or am I doing
something wrong?) and switchApplicationLanguage has just been made
(see kdelibs) and as such the action does nothing but does not
cause trouble either so - it has been added.

TODO: Tips of the day