kdeplasma-addons: check for MPRIS interface instead of relying on categories
generally speaking trying to guess MPRIS interfaces is brute-force guess - it
works kinda. the check that was used before this change however does not
account for applications that are not categorized as audio or music player
but do have MPRIS interface (chromium does).

side note: tested only the MPRIS v2 interface

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>